Fix narrowing type casts in Fully Connected operator
PiperOrigin-RevId: 306765979

+++ src/fully-connected-nc.c
@@ -132,8 +132,8 @@ enum xnn_status xnn_create_fully_connected_nc_q8(
   const uint32_t nr = xnn_params.q8.gemm.nr;
   const uint32_t kr = UINT32_C(1) << xnn_params.q8.gemm.log2_kr;
 
-  const uint32_t n_stride = round_up(output_channels, nr);
-  const uint32_t k_stride = round_up_po2(input_channels, kr);
+  const size_t n_stride = round_up(output_channels, nr);
+  const size_t k_stride = round_up_po2(input_channels, kr);
 
   fully_connected_op->packed_weights = xnn_allocate_simd_memory(n_stride * (k_stride * sizeof(uint8_t) + sizeof(int32_t)));
   if (fully_connected_op->packed_weights == NULL) {
@@ -274,8 +274,8 @@ enum xnn_status xnn_create_fully_connected_nc_f32(
   const uint32_t kr = UINT32_C(1) << xnn_params.f32.gemm.log2_kr;
   const uint32_t sr = UINT32_C(1) << xnn_params.f32.gemm.log2_sr;
 
-  const uint32_t n_stride = round_up(output_channels, nr);
-  const uint32_t k_stride = round_up_po2(input_channels, kr);
+  const size_t n_stride = round_up(output_channels, nr);
+  const size_t k_stride = round_up_po2(input_channels, kr);
 
   fully_connected_op->packed_weights = xnn_allocate_simd_memory(n_stride * (k_stride * sizeof(float) + sizeof(float)));
   if (fully_connected_op->packed_weights == NULL) {
